version 1.0 done

+++ server.js
@@ -20,6 +20,21 @@ let tasks = [
   }
 ];
 
+const getByIdRequest = (req, res) =>
+{
+  const id = req.url.split('/')[2];
+  const task = tasks.find(task => task.id === parseInt(id));
+
+  res.writeHead(200, { 'content-type': 'application/json' });
+  res.end(JSON.stringify(task));
+};
+
+const getrequest = (req, res) => 
+{
+  res.writeHead(200, { 'content-type': 'application/json' })
+  res.end(JSON.stringify(tasks))
+};
+
 const postRequest = (req, res) =>
 {
   let body = '';
@@ -34,22 +49,24 @@ const postRequest = (req, res) =>
     {
       const parsedData = JSON.parse(body);
 
-      console.log("parsed Data:", parsedData);
+      const newTask = { id: tasks.length + 1, task: parsedData.task, complete: parsedData.complete};
+      console.log("parsed Data:", newTask);
 
-      if(typeof parsedData.id !== 'number' || 
-        typeof parsedData.task !== 'string' ||
+      if(typeof parsedData.task !== 'string' ||
         typeof parsedData.complete !== 'boolean') 
       {
         throw new Error("invalid data types");
-      } else
+      } else if (req.headers['content-type'] !== 'application/json')
       {
-        console.log("tasks array before POST:", tasks);
-        tasks.push(parsedData);
-        console.log("tasks array after POST:", tasks);
+        res.writeHead(415);
+        return res.end(JSON.stringify({ error: "Content-Type must be application/json" }))
+      }
+      console.log("tasks array before POST:", tasks);
+      tasks.push(newTask);
+      console.log("tasks array after POST:", tasks);
 
-        res.writeHead(201, { 'Content-Type': 'application/json' });
-        res.end(JSON.stringify(parsedData));
-      };
+      res.writeHead(201, { 'Content-Type': 'application/json' });
+      res.end(JSON.stringify(parsedData));
     } catch (error)
     {
       res.writeHead(400);
@@ -60,32 +77,86 @@ const postRequest = (req, res) =>
 
 const deleteItemById = (req, res) => 
 {
-  tasks = newTasks;
-  console.log(newTasks);
-  res.end();
-}
+  const id = Number(req.url.split('/')[2]);
+  const taskIndex = tasks.findIndex(task => task.id === parseInt(id));
+
+  if(taskIndex === -1)
+  {
+    res.writeHead(404, { 'Content-Type': 'application/json' })
+    res.end("Item ID Not Found");
+  } else 
+  {
+    tasks.splice(taskIndex, 1);
+    console.log(tasks);
+    
+    res.writeHead(200, { 'Content-Type': 'application/json' })
+    res.end(JSON.stringify(tasks));
+  };
+};
+
+const putRequest = (req, res) =>
+{
+  let body = '';
+  
+  req.on('data', chunk => {
+    body += chunk.toString();
+  });
+  
+  req.on('end', () =>
+  {
+    try
+    {
+      const parsedData = JSON.parse(body); 
+      const id = Number(req.url.split('/')[2]);
+      const task = tasks.find(task => task.id === parseInt(id));
+
+      if(typeof parsedData.task !== 'string' ||
+        typeof parsedData.complete !== 'boolean') 
+      {
+        throw new Error("invalid data types");
+      } else if (req.headers['content-type'] !== 'application/json')
+      {
+        res.writeHead(415);
+        return res.end(JSON.stringify({ error: "Content-Type must be application/json" }))
+      };
+      
+      task.task = parsedData.task;
+      task.complete = parsedData.complete;
+      
+      console.log("task after update:", task);
+
+      console.log("tasks after update:", tasks);
+      res.writeHead(200)
+      res.end('array has been updated successfully')
+    } catch (error)
+    {
+      res.writeHead(400)
+      res.end(JSON.stringify({ error: error.message }))
+    };
+  });
+};
 
 const server = http.createServer((req, res) =>
 {
   if(req.method === 'GET' && req.url === '/tasks')
   {
-    const allTasks = JSON.stringify(tasks);
-    res.writeHead(200, {'content-type': 'application/json'});
-    res.end(allTasks);
+    getrequest(req, res);
+  }else if(req.method === 'GET' && req.url.match(/\/tasks\/([0-9]+)/))
+  {
+    getByIdRequest(req, res);
   } else if(req.method === 'POST' && req.url === '/tasks') 
   {
     postRequest(req, res);
-    // console.log(tasks);
-  } else if (req.method === 'DELETE')
+  } else if (req.method === 'DELETE' && req.url.match(/\/tasks\/([0-9]+)/))
   {
     deleteItemById(req, res);
-  } else if (req.method === 'PUT')
+  } else if (req.method === 'PUT' && req.url.match(/\/tasks\/([0-9]+)/))
   {
-
+    putRequest(req, res);
   } else 
   {
     res.writeHead(400, {'content-type' : 'application/json'});
-    res.end('ERROR');
+    res.end('invalid URL');
   };
 });
